# openapi.yaml
openapi: 3.0.3
info:
  title: TSTS - Technical Support Ticketing System API
  description: |
    Comprehensive API documentation for the Technical Support Ticketing System (TSTS).
    
    ## Overview
    TSTS is a full-featured ticketing system designed for university IT support operations.
    It includes ticket management, chat, knowledge base, AI assistance, and comprehensive reporting.
    
    ## Authentication
    The API supports multiple authentication methods:
    - **Bearer Token**: JWT-based authentication for web and mobile clients
    - **API Keys**: For third-party integrations with scope-based access control
    - **WebAuthn**: Passwordless authentication using trusted devices
    
    ## Base URLs
    - Production: `https://api.tsts.local`
    - Development: `http://localhost:3000`
    
    ## Rate Limiting
    - 1000 requests per 15 minutes per IP address
    - Rate limit headers included in responses
    
    ## Pagination
    Most list endpoints support pagination with `page` and `limit` query parameters:
    - Default page: 1
    - Default limit: 20
    - Maximum limit: 200
    
    ## Internationalization
    The API supports English and Arabic. Many fields use JSONB format:
    ```json
    {
      "name": {
        "en": "Computer Science",
        "ar": "علوم الحاسب"
      }
    }
    ```
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "message": "Error description",
      "statusCode": 400,
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party integrations
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v1/users/requesters:
    get:
      summary: List requesters
      tags: [Users]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of requesters

    post:
      summary: Create requester
      tags: [Users]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, email, ssn, departmentId, universityId]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                ssn:
                  type: string
                departmentId:
                  type: string
                  format: uuid
                universityId:
                  type: string
                  format: uuid
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Requester created
...